+++ Cargo.toml
@@ -5,5 +5,4 @@ members = [
     "crates/macros",
     "crates/ai-tools",
     "crates/shared",
-    "crates/tools",
-    "crates/tool-cat", "crates/tool-csvn", "crates/tool-eventhub-export", "crates/tool-eventhub-read", "crates/tool-get-lines", "crates/tool-guid", "crates/tool-jwt", "crates/tool-split", "crates/tool-touch"]
+    "crates/tool-cat", "crates/tool-csvn", "crates/tool-eventhub-export", "crates/tool-eventhub-read", "crates/tool-get-lines", "crates/tool-guid", "crates/tool-jwt", "crates/tool-split", "crates/tool-touch", "crates/tool-timestamp"]
