+++ crates/tool-timestamp/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "ts"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+shared = { path = "../shared" }
+clap = "4.5.45"
+chrono = "0.4.41"
+anyhow = "1.0.99"
+tracing = "0.1.41"
