+++ crates/tool-timestamp/src/models.rs
