--- crates/tools/Cargo.toml
@@ -1,68 +1,0 @@
-[package]
-name = "rusted-toolbox"
-version = "1.0.0"
-edition = "2021"
-authors = ["Breno RdV"]
-description = "A collection of CLI utilities ported from Go, Python, PowerShell, and C# to Rust. Plus a few more."
-
-
-[[bin]]
-name = "ts"
-path = "src/bin/ts.rs"
-
-
-[dependencies]
-shared = { path = "../shared" }
-clap = { version = "4.5.45", features = ["derive"] }
-dotenv = "0.15.0"
-tokio = { version = "1.47", features = ["full"] }
-uuid = { version = "1.18", features = ["v4"] }
-arboard = "3.6"
-csv = "1.3"
-chrono = { version = "0.4", features = ["serde"] }
-anyhow = "1.0.99"
-thiserror = "1.0"
-serde = { version = "1.0", features = ["derive"] }
-regex = "1.11.1"
-futures = "0.3"
-async-std = { version = "1.13", features = ["attributes"] }
-encoding_rs = "0.8"
-num_cpus = "1.17"
-filetime = "0.2"
-jsonwebtoken = "9.3"
-serde_json = "1.0"
-colored = "2.2"
-base64 = "0.21"
-ctrlc = "3.4"
-sled = "0.34"
-crossterm = "0.27"
-indicatif = "0.17"
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-fastrand = "2.3"
-log = "0.4"
-azeventhubs = "0.20"
-azure_core = "0.20"
-azure_identity = "0.20"
-bytes = "1.10"
-url = "2.5.4"
-futures-util = "0.3"
-async-trait = "0.1.89"
-tokio-stream = "0.1.17"
-once_cell = "1.21.3"
-memmap2="0.9.7"
-string-interner="0.19.0"
-reqwest = { version = "0.12.23", features = ["json"] }
-dialoguer = "0.11.0"
-notify="8.2.0"
-decompress = "0.6.0"
-rusqlite = { version = "0.37.0", features = ["bundled"] }
-rusted_toolbox_macros = { path = "../macros" }
-walkdir = "2.5.0"
-tracing-log = "0.2.0"
-infer="0.19.0"
-
-
-[dev-dependencies]
-tempfile = "3.20.0"
-rstest="0.26.1"
