--- crates/tools/src/tools/mod.rs
@@ -1,1 +1,0 @@
-pub mod ts;
