--- crates/tools/src/tools/ts/mod.rs
@@ -1,3 +1,0 @@
-pub mod cli_utils;
-pub mod models;
-pub mod ts_app;
